Modify Views structure
Decompose Admin View :
- Conf components
- File manager components
