+++ src/services/folder/folder.service.js
@@ -5,34 +5,27 @@ const API_URL = 'http://localhost:8080/blog/folders/';
 
 class FolderService{
 
+	getRootFolder(){
+		return axios
+			.get(API_URL + 'root/', { headers: authHeader()})
+			.then(response => { return response.data; });
+	}
+
 	getFolder(path){
 		return axios
-			.get(API_URL + "?path=" + path, 
-			{
-				headers: authHeader()
-			}
-
-		).then(response => {
-			return response.data;
-		});
+			.get(API_URL + "?path=" + path, {headers: authHeader()})
+			.then(response => {	return response.data; });
 	}
 
 	addFolder(folder, parentPath){
 		return axios
-			.post(API_URL + "?parentPath=" + parentPath, folder, 
-			{
-				headers: authHeader()
-			}
+			.post(API_URL + "?parentPath=" + parentPath, folder, {headers: authHeader()}
 		);
 	}
 
 	deleteFolder(path){
 		return axios
-			.delete(API_URL + "?path=" + path,
-			{
-				headers: authHeader()
-			}
-		);
+			.delete(API_URL + "?path=" + path, {headers: authHeader()});
 	}
 
 }
